+++ server/playerpool.cpp
@@ -313,7 +313,7 @@ bool CPlayerPool::IsNickInUse(PCHAR szNick)
 {
 	int x=0;
 	while(x!=MAX_PLAYERS) {
-		if(GetSlotState((BYTE)x)) {
+		if(GetSlotState(x)) {
 			//if(!stricmp(GetPlayerName((BYTE)x),szNick)) {
 			if (!strcmp(m_pPlayers[x]->GetName(), szNick)) {
 				return true;
